requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
alembic
pydantic-settings
python-dotenv
beautifulsoup4
requests
pyyaml
pdfplumber
anthropic
playwright>=1.48.0
python-docx>=1.1.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
